# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  invelog_pkg_models.Category:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  invelog_pkg_models.Container:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      location:
        $ref: '#/definitions/invelog_pkg_models.Location'
      location_id:
        type: string
      name:
        type: string
      parent:
        $ref: '#/definitions/invelog_pkg_models.Container'
      parent_id:
        type: string
      project:
        $ref: '#/definitions/invelog_pkg_models.Project'
      project_id:
        type: string
      updated_at:
        type: string
    type: object
  invelog_pkg_models.Item:
    properties:
      category:
        $ref: '#/definitions/invelog_pkg_models.Category'
      category_id:
        type: string
      checked_out:
        type: boolean
      container:
        $ref: '#/definitions/invelog_pkg_models.Container'
      container_id:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: string
      individual_notes:
        type: string
      item_type:
        $ref: '#/definitions/invelog_pkg_models.ItemType'
      item_type_id:
        type: string
      last_check_out_time:
        type: string
      name:
        description: Used if ItemType is not provided
        type: string
      origin_location:
        $ref: '#/definitions/invelog_pkg_models.Location'
      origin_location_id:
        type: string
      quantity:
        type: integer
      serial_number:
        type: string
      updated_at:
        type: string
    type: object
  invelog_pkg_models.ItemType:
    properties:
      category:
        $ref: '#/definitions/invelog_pkg_models.Category'
      category_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      manufacturer:
        type: string
      name:
        type: string
      part_number:
        type: string
      specifications:
        type: string
      updated_at:
        type: string
    type: object
  invelog_pkg_models.Location:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  invelog_pkg_models.Project:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      status:
        description: e.g., "active", "completed", "archived"
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive inventory management system backend for electronic
    components.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Invelog API
  version: "1.0"
paths:
...
  /categories:
    get:
      description: Get all categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/invelog_pkg_models.Category'
            type: array
      summary: List Categories
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Create a new category
      parameters:
      - description: Category Data
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/invelog_pkg_models.Category'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/invelog_pkg_models.Category'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create Category
      tags:
      - Categories
  /categories/{id}:
    delete:
      description: Delete a category by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete Category
      tags:
      - Categories
    get:
      description: Get a category by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/invelog_pkg_models.Category'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Category
      tags:
      - Categories
    put:
      consumes:
      - application/json
      description: Update a category by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Category Data
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/invelog_pkg_models.Category'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/invelog_pkg_models.Category'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update Category
      tags:
      - Categories
  /containers:
    get:
      description: Get all containers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/invelog_pkg_models.Container'
            type: array
      summary: List Containers
      tags:
      - Containers
    post:
      consumes:
      - application/json
      description: Create a new container
      parameters:
      - description: Container Data
        in: body
        name: container
        required: true
        schema:
          $ref: '#/definitions/invelog_pkg_models.Container'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/invelog_pkg_models.Container'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create Container
      tags:
      - Containers
  /containers/{id}:
    delete:
      description: Delete a container by ID
      parameters:
      - description: Container ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete Container
      tags:
      - Containers
    get:
      description: Get a container by ID
      parameters:
      - description: Container ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/invelog_pkg_models.Container'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Container
      tags:
      - Containers
    put:
      consumes:
      - application/json
      description: Update a container by ID
      parameters:
      - description: Container ID
        in: path
        name: id
        required: true
        type: string
      - description: Container Data
        in: body
        name: container
        required: true
        schema:
          $ref: '#/definitions/invelog_pkg_models.Container'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/invelog_pkg_models.Container'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update Container
      tags:
      - Containers
...
  /items/{id}/checkin:
    post:
      description: Check in an item by marking it as not checked out and logging the
        activity
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/invelog_pkg_models.Item'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Check in an item
      tags:
      - Items
  /items/{id}/checkout:
    post:
      description: Check out an item by marking it as checked out and logging the
        activity
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/invelog_pkg_models.Item'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Check out an item
      tags:
      - Items
...
swagger: "2.0"